Fixed bug that prevented file saving
### Description:
The previous commit had introduced a bug that prevented the FileSystemHandler from saving files due to a DirectoryNotFound error.
This was caused by me mistakenly adding the FileSystemHandler's TimestampFormat in the SessionManager.SessionTitle variable. But because the session title is used to name the saved file and the aforementioned timestamp format utilized colons (":"), the system failed to create the file due to colons being an illegal character in file names.
This bug has now been fixed.