Implemented DeviceModeChecker to auto-enable XR / non-XR char controller
The DeviceModeChecker class triggers the OnModeMatch and OnModeMismatch UnityEvents based on the selected DeviceMode.

Example: If the global Settings.deviceMode variable is set to "XR", any DeviceModeChecker set to "Desktop" will trigger the OnModeMismatch event. On the contrary, any DeviceModeChecker set to "XR" will trigger the OnModeMatch event. These events may be hooked up to the SetActive() methods of the XR and Desktop player character to handle automatic enabling / disabling of them based on the global setting value.