fix: Updated DOM to not auto-assign IDs during play mode anymore
As any changes during play mode, including the auto-assigned ID of the DOM, are cleared once exiting play mode, this feature simply did not work. IDs have to be generated before entering play mode, but that can be done by bulk-selecting every DynamicObject that does not have an ID yet and pressing the "Request ID" button in the inspector.