fix: Creation of local GPs does not create additional global GPs anymore
### Explanation:
Before this fix, whenever the GazePointManager created a local GazePoint (aka a GazePoint with a DynamicObject reference), it also created a global GazePoint equivalent in the same frame. This was simply due to me having forgotten to put in the necessary "return" in one line.
However, even with this fixed, the implementation was still more hacked together to accomodate for initially unaddressed cases than actually planned out. Therefore, I reworked the implementation to now utilize the DynamicObjectColliderHelper class for any DynamicObject that carries its colliders on a child object. This child is simply equipped with the Helper class, allowing me to forgo the parent check in the GPM, which would have otherwise been executed at a rate equal to tickrate. This approach additionally allows door handles to count as part of the DynamicObject door as well.