refactor: Changed the GameStart distance check to run in coroutine
As this was a very severe change to the class, its name has been updated as well.

### Disclaimer:
Now, the TriggerBasedOnPlayerDistance object is intended to be enabled on application start. This way, events are used to tell the object to start / stop checking the distance of the player.

Also, it only checks for the distance on the x-z plane now. Vertical distance of the player to the "starting area" is ignored entirely, resulting in a cylindrical "hitbox"